Updated star wrapper.

+++ STAR/runSTARBatch.py
@@ -24,8 +24,8 @@ from multiprocessing import Pool as ThreadPool
 #     if os.path.isfile(fname):
 #         listOfPrefixes = [x for x in listOfPrefixes if x != i]
 # print(len(listOfPrefixes))
-fastqGzFiles = glob.glob("*.fastq.gz")
-fastqGzPairs = f.pairGivenFastqFiles(fastqGzFiles)
+# fastqGzFiles = glob.glob("*.fastq.gz")
+# fastqGzPairs = f.pairGivenFastqFiles(fastqGzFiles)
 
 # pairFiles = []
 # singleFiles = []
@@ -38,11 +38,11 @@ fastqGzPairs = f.pairGivenFastqFiles(fastqGzFiles)
 
 import time
 
-def runStarPairWrap(tupleOfFiles, numthreads = 16):
+def runStarPairWrap(tupleOfFiles, genome, numthreads = 16):
     print("\n\n")
     startTime = time.time()
     # print("Starting STAR run on " + str(tupleOfFiles))
-    result = star.runStar(tupleOfFiles[0], tupleOfFiles[1], "/media/rawData/genomes/STAR_genomeDir_hg19_vGATK", cpu = numthreads)
+    result = star.runStar(tupleOfFiles[0], tupleOfFiles[1], genome, cpu = numthreads)
     if result != None:
         logfile = open(f.longestCommonSubstring(tupleOfFiles[0], tupleOfFiles[1]) + '.star.log', 'w')
         logfile.write(result)
@@ -62,24 +62,45 @@ def runStarSingleWrap(file, numthreads = 12):
 
 # Parallel(n_jobs=2)(delayed(runStarWrap)(i) for i in listOfFiles)
 
-for x in fastqGzPairs:
-    extractedFiles = (x[0][:len(x[0]) - 3], x[1][:len(x[1]) - 3])
-    if os.path.isfile(extractedFiles[0]) != True or os.path.isfile(extractedFiles[1]) != True:
-        extractCommand = "zcat %s > %s"
-        extractCommandList = []
-        extractCommandList.append(extractCommand % (x[0], extractedFiles[0]))
-        extractCommandList.append(extractCommand % (x[1], extractedFiles[1]))
-        pool = ThreadPool(2)
-        print("Extracting fastq file pair...")
-        pool.map(s.executeFunctions,extractCommandList)
-    else:
-        print("Fastq files already extracted.")
-    # s.executeFunctions(extractCommand % (x[0], x[0]), captureOutput=False)
-    # s.executeFunctions(extractCommand % (x[1], x[1]), captureOutput=False)
-    print("Running STAR...")
-    runStarPairWrap(extractedFiles)
-    s.executeFunctions("rm *.fastq")
+def main():
+    try:
+        optlist, args = getopt.getopt(args = sys.argv[1:], shortopts=None, longopts = [
+            'genomeDir='])
+    except getopt.GetoptError as err:
+        print(err)
+        sys.exit(2)
+    genomeDir = ""
+    for o, a in optlist:
+        if o == '--genomeDir':
+            genomeDir = a
+    if len(args) == 0:
+        print("No fastq files given. Exiting.")
+        sys.exit(2)
+    if len(genomeDir) == 0:
+        print("No genome directory specified. Exiting.")
+        sys.exit(2)
 
+    for x in args:
+        extractedFiles = (x[0][:len(x[0]) - 3], x[1][:len(x[1]) - 3])
+        if os.path.isfile(extractedFiles[0]) != True or os.path.isfile(extractedFiles[1]) != True:
+            extractCommand = "zcat %s > %s"
+            extractCommandList = []
+            extractCommandList.append(extractCommand % (x[0], extractedFiles[0]))
+            extractCommandList.append(extractCommand % (x[1], extractedFiles[1]))
+            pool = ThreadPool(2)
+            print("Extracting fastq file pair...")
+            pool.map(s.executeFunctions,extractCommandList)
+        else:
+            print("Fastq files already extracted.")
+        # s.executeFunctions(extractCommand % (x[0], x[0]), captureOutput=False)
+        # s.executeFunctions(extractCommand % (x[1], x[1]), captureOutput=False)
+        print("Running STAR...")
+        runStarPairWrap(extractedFiles, genomeDir)
+        s.executeFunctions("rm *.fastq")
+    star.unloadGenome(genomeDir)
+
+if __name__ == '__main__':
+    main()
 # for x in singleFiles:
 #     extractCommand = "zcat %s.gz > %s" % (x, x)
 #     print("Extracting fastq files...")
